requirements.txt
pytaglib
peewee
pygame
PyQt5